move well-known kubelet cloud provider annotations to k8s.io/cloud-provider
