+++ pkg/kubelet/apis/BUILD
@@ -8,7 +8,6 @@ load(
 go_library(
     name = "go_default_library",
     srcs = [
-        "well_known_annotations.go",
         "well_known_annotations_windows.go",
         "well_known_labels.go",
     ],
